Change memcmplen parameters and result to size_t

+++ src/liblzma/common/memcmplen.h
@@ -39,9 +39,9 @@
 ///             It's rounded up to 2^n. This extra amount needs to be
 ///             allocated in the buffers being used. It needs to be
 ///             initialized too to keep Valgrind quiet.
-static inline uint32_t lzma_attribute((__always_inline__))
+static inline size_t lzma_attribute((__always_inline__))
 lzma_memcmplen(const uint8_t *buf1, const uint8_t *buf2,
-		uint32_t len, uint32_t limit)
+		size_t len, size_t limit)
 {
 	assert(len <= limit);
 	assert(limit <= UINT32_MAX / 2);
